Added Websockets and setup a basic Chat App
Configured settings and asgi.py to work with the channels. Modified Url
routing for the new app. Setup up views, models, consumers and routing
for the app. Worked on templates and added neccessart js for the socket
to work

This Chat App was setup to understand the basics and working of Django
Channels and Web Sockets